services/eval-service/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
openai-whisper==20231117
anthropic==0.25.0
torch==2.2.2
torchaudio==2.2.2
scipy==1.13.0
soundfile==0.12.1
python-dotenv==1.2.2